Moved Properties from iidm to stdcxx
