+++ test/iidm/CMakeLists.txt
@@ -27,7 +27,6 @@ set(UNIT_TEST_SOURCES
     NetworkTest.cpp
     NodeBreakerVoltageLevelTest.cpp
     PhaseTapChangerTest.cpp
-    PropertiesTest.cpp
     RatioTapChangerTest.cpp
     ReactiveLimitsTest.cpp
     ShuntCompensatorTest.cpp
